[Category] Button
[Issue] The 'OK' button is displayed with a wrong animation.

[File] garden/bg.swf

[Steps to reproduce]
1. Play back the file.
2. The 'OK' button appears at the bottom, with a blink animation.

[Expected behavior]
The animation is performed by switching two images, one is the button with shadow, the other is the non-shadow.

[Actual result]
Same as above, except that the shadow is displayed in a wrong position.

[Cause]
Transform matrix is not applied when the shadowed button is added to the stage.
There are two code paths for adding display objects; actions.add() and actions.replace()
The shadowed button is added by action.replace() while non-shadowed button is by action.add()
The issue is that the code applying transform matrix is missing from action.replace()

[Fix]
Added the code applying transform matrix to the action.replace()